+++ docs/STEP2_REQUIREMENTS.md
@@ -1,0 +1,32 @@
+Tap <br/>
+
+- [ ] 탭을 추가한다.
+- [ ] 모든 음식점을 클릭하면 모든 음식점이 나온다.
+- [ ] 원래 필터의 선택값이 유지된다.
+- [ ] 자주 가는 음식점을 클릭하면 favoriteList가 나온다.
+
+RestaurantList
+
+- [ ] 선택한 음식점이 없을 경우 메세지를 보여준다(UI 관련)
+
+FavoriteIcon
+
+- [ ] 음식점 Item에서 별을 클릭하면 favorite 속성이 반대로 바뀐다.
+
+DetailModal
+
+- [ ] 카테고리 아이콘, 제목, 거리, 디테일, 링크가 뜬다.
+- [ ] FavoriteIcon을 포함한다.
+
+- [ ] 삭제하기 버튼을 누르면 '정말 삭제하시겠습니까?' 하는 알림으로 삭제 확인을 받는다.
+
+- [ ] 삭제 확인에서 yes를 입력하면
+- [ ] 해당 데이터가 삭제된다.
+- [ ] 모달이 없어진다.
+
+- [ ] 삭제 확인에서 no를 입력하면
+- [ ] 원래 디테일 모달로 돌아간다.
+
+- [ ] 닫기 버튼을 누르면 modal이 꺼진다.
+
+e2e 테스트
