+++ src/components/BasicModal/BasicModal.ts
@@ -1,6 +1,7 @@
-import { blockModalBodyScroll } from '@/utils/view';
+import { blockModalBodyScroll, closeModal } from '@/utils/view';
 import BaseComponent from '../BaseComponent';
 import { $ } from '@/utils/DOM';
+import { ErrorMessage } from '@/constants/Message';
 
 class BasicModal extends BaseComponent {
   #children: Node;
@@ -18,9 +19,10 @@ class BasicModal extends BaseComponent {
 
   setEvent(): void {
     this.#backdropElement.addEventListener('click', () => {
-      if ($('.modal')) {
-        blockModalBodyScroll();
-      }
+      const $modal = $('.modal');
+      if (!$modal) return console.error(ErrorMessage.NULL_SELECTOR);
+      closeModal($modal);
+      blockModalBodyScroll();
     });
   }
